Spies Update
Cleaned up some of the code on the spies page, moving some of the
methods over to Spy Scripts.  Found and issue with Captcha, so had to
change some of how that worked.  Added a method to the account manager
that returns a bool as to whether or not a captcha is valid.  This check
should be run before any attempt to use a method that requires a
captcha.  Also added some missing pieces to the Spies Response Wrapper
about how many offensive or defensive missions have been run.